render youtube, simplify
