+++ lib/types/image.js
@@ -19,8 +19,4 @@ export const parse = ([elm]) => {
 };
 
 export const render = ({type, src, alt, width, height}) =>
-  type === 'image'
-  ? h('img', {
-    src, alt, width, height
-  })
-  : null;
+  h('img', { src, alt, width, height });
