+++ lib/types/video.js
@@ -38,6 +38,4 @@ export const parse = ([elm]) => {
 const renderSource = ({src, type}) => h('source', {src, type});
 
 export const render = ({type, sources, width, height}) =>
-  type === 'video'
-  ? h('video', {width, height}, sources.map(renderSource))
-  : null
+  h('video', {width, height}, sources.map(renderSource));
