+++ lib/types/youtube.js
@@ -1,5 +1,6 @@
 import { parse as parseEmbedly } from 'embedly-url';
 import getYoutubeId from 'get-youtube-id';
+import {element as h} from 'deku';
 
 export const parse = ([elm]) => {
   let tagName = elm.tagName.toLowerCase();
@@ -21,3 +22,6 @@ export const parse = ([elm]) => {
     youtubeId
   };
 };
+
+export const render = ({youtubeId}) =>
+  h('iframe', {src: `https://www.youtube.com/embed/${youtubeId}`});
