+++ test.js
@@ -224,6 +224,15 @@ test('parse() youtube embedly iframe', t => {
   t.same(actual, expected);
 });
 
+test('render() youtube iframe', t => {
+  const actual = render({
+    type: 'youtube',
+    youtubeId: 'pDVmldTurqk'
+  });
+  const expected = '<iframe src="https://www.youtube.com/embed/pDVmldTurqk"></iframe>';
+  t.same(actual, expected);
+});
+
 test('parse() tweet - normal', t => {
   const input = `<blockquote class="twitter-tweet" lang="en"><p lang="en" dir="ltr">GIF vs. JIF… This <a href="https://t.co/qFAHWgdbL6">pic.twitter.com/qFAHWgdbL6</a></p>&mdash; Matt (foo) Navarra (@MattNavarra) <a href="https://twitter.com/MattNavarra/status/684690494841028608">January 6, 2016</a></blockquote><script async src="//platform.twitter.com/widgets.js" charset="utf-8"></script>`;
   const actual = parse(input);
